requirements.txt
kagglehub
streamlit
pandas-profiling
seaborn
scikit-learn